requirements.txt
pyautogui
pillow
tkcap
pydicom
img2pdf
opencv_python
